fix: doi cot calories trong dishes sang decimal de luu dung so le VDD
Truoc gio calories la integer nen bi lam tron (vd 420.5 -> 420 hoac
421), sai lech voi so lieu goc tu Vien Dinh duong. Doi sang
numeric(8,2) + cast float trong model, cap nhat validation admin va
input UI cho phep nhap so le. Reseed lai VddDishSeeder voi fixture
giu nguyen 1 chu so thap phan tu API.

+++ app/Models/Dish.php
@@ -23,7 +23,7 @@ class Dish extends Model
 
     protected $casts = [
         'aliases'  => 'array',
-        'calories' => 'integer',
+        'calories' => 'float',
         'protein'  => 'integer',
         'carbs'    => 'integer',
         'fat'      => 'integer',
